pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Implementation of the challenge for the company Credito Express"
authors = ["Rafael Cascalho <cascalhoeng@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
python-dotenv = "^0.15.0"
uvicorn = "^0.13.4"
motor = "^2.3.1"
PyJWT = "^2.0.1"
http3 = "^0.6.7"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
seeder = "scripts:seeder"
server = "scripts:web_server"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
